backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
requests==2.28.2
numpy==1.24.2
llama-cpp-python==0.1.77
sentence-transformers==2.2.2
torch==2.0.1
beautifulsoup4==4.12.2